Move LSP client management code to its own class
