+++ src/godot-tools.ts
@@ -3,43 +3,29 @@ import * as path from "path";
 import * as vscode from "vscode";
 import { GDDocumentLinkProvider } from "./document_link_provider";
 import GDScriptLanguageClient, { ClientStatus } from "./lsp/GDScriptLanguageClient";
+import { ClientConnectionManager } from "./lsp/ClientConnectionManager";
 import { ScenePreviewProvider } from "./scene_preview_provider";
 import { get_configuration, set_configuration, get_godot_version, get_project_dir, find_file, set_context, find_project_file } from "./utils";
 
 const TOOL_NAME = "GodotTools";
 
 export class GodotTools {
-	private reconnection_attempts = 0;
-
 	private context: vscode.ExtensionContext;
-	private client: GDScriptLanguageClient = null;
+
+	private lspClientManager: ClientConnectionManager = null;
 	private linkProvider: GDDocumentLinkProvider = null;
 	private scenePreviewManager: ScenePreviewProvider = null;
 
-	private connection_status: vscode.StatusBarItem = null;
-
 	constructor(p_context: vscode.ExtensionContext) {
 		this.context = p_context;
-		this.client = new GDScriptLanguageClient(p_context);
-		this.client.watch_status(this.on_client_status_changed.bind(this));
-		this.connection_status = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right);
 
+		this.lspClientManager = new ClientConnectionManager(p_context);
 		this.linkProvider = new GDDocumentLinkProvider(p_context);
-
-		setInterval(() => {
-			this.retry_callback();
-		}, get_configuration("lsp.autoReconnect.cooldown", 3000));
 	}
 
 	public activate() {
 		vscode.commands.registerCommand("godotTools.openEditor", () => {
 			this.open_workspace_with_editor("-e").catch(err => vscode.window.showErrorMessage(err));
-		});
-		vscode.commands.registerCommand("godotTools.startLanguageServer", () => {
-			this.start_language_server().catch(err => vscode.window.showErrorMessage(err));
-		});
-		vscode.commands.registerCommand("godotTools.stopLanguageServer", () => {
-			this.stop_language_server();
 		});
 		vscode.commands.registerCommand("godotTools.runProject", () => {
 			this.open_workspace_with_editor().catch(err => vscode.window.showErrorMessage(err));
@@ -47,77 +33,17 @@ export class GodotTools {
 		vscode.commands.registerCommand("godotTools.runProjectDebug", () => {
 			this.open_workspace_with_editor("--debug-collisions --debug-navigation").catch(err => vscode.window.showErrorMessage(err));
 		});
-		vscode.commands.registerCommand("godotTools.checkStatus", this.check_client_status.bind(this));
 		vscode.commands.registerCommand("godotTools.setSceneFile", this.set_scene_file.bind(this));
 		vscode.commands.registerCommand("godotTools.copyResourcePathContext", this.copy_resource_path.bind(this));
 		vscode.commands.registerCommand("godotTools.copyResourcePath", this.copy_resource_path.bind(this));
 		vscode.commands.registerCommand("godotTools.openTypeDocumentation", this.open_type_documentation.bind(this));
 		vscode.commands.registerCommand("godotTools.switchSceneScript", this.switch_scene_script.bind(this));
 
-		set_context("godotTools.context.connectedToEditor", false);
-
 		this.scenePreviewManager = new ScenePreviewProvider();
-
-		this.connection_status.text = "$(sync) Initializing";
-		this.connection_status.command = "godotTools.checkStatus";
-		this.connection_status.show();
-
-		this.connect_to_language_server()
 	}
 
 	public deactivate() {
-		this.client.stop();
-	}
-
-	// TODO: move to LSP client?
-	private connect_to_language_server() {
-		const start = get_configuration("lsp.runAtStartup", false)
-
-		if (start) {
-			this.start_language_server();
-		}
-
-		this.reconnection_attempts = 0;
-		this.client.connect_to_server();
-	}
-
-	// TODO: move to LSP client?
-	private stop_language_server() {
-		const existingTerminal = vscode.window.terminals.find(t => t.name === `${TOOL_NAME}LSP`);
-		if (existingTerminal) {
-			existingTerminal.dispose();
-		}
-	}
-
-	// TODO: move to LSP client?
-	private start_language_server() {
-		return new Promise<void>(async (resolve, reject) => {
-			const project_dir = await get_project_dir();
-
-			if (!project_dir) {
-				reject("Current workspace is not a Godot project");
-			}
-
-			const godot_version = await get_godot_version();
-
-			let headlessFlag = "--no-window";
-			if (godot_version.startsWith('4')) {
-				headlessFlag = "--headless";
-			}
-
-			// TODO: find a better way to manage child processes
-			// This way works, but it creates a terminal that the user might
-			// accidentally close, and start a confusing cycle.
-			// I also want to be able to monitor the child process for errors
-			// or exits, and restart it as appropriate.
-
-			this.stop_language_server();
-
-			const editorPath = get_configuration("editorPath", "");
-			const command = `${editorPath} --path "${project_dir}" --editor ${headlessFlag}`;
-			const terminal = vscode.window.createTerminal(`${TOOL_NAME}LSP`);
-			terminal.sendText(command, true);
-		});
+		this.lspClientManager.client.stop();
 	}
 
 	private open_workspace_with_editor(params = "") {
@@ -170,7 +96,7 @@ export class GodotTools {
 		const wordRange = document.getWordRangeAtPosition(curPos);
 		const symbolName = document.getText(wordRange);
 
-		this.client.open_documentation(symbolName);
+		this.lspClientManager.client.open_documentation(symbolName);
 	}
 
 	private async switch_scene_script() {
@@ -282,100 +208,6 @@ export class GodotTools {
 				});
 			} else {
 				run_godot(editorPath, params);
-			}
-		});
-	}
-
-	// TODO: move to LSP client?
-	private check_client_status() {
-		let host = get_configuration("lsp.serverPort", "localhost");
-		let port = get_configuration("lsp.serverHost", 6008);
-		switch (this.client.status) {
-			case ClientStatus.PENDING:
-				vscode.window.showInformationMessage(`Connecting to the GDScript language server at ${host}:${port}`);
-				break;
-			case ClientStatus.CONNECTED:
-				vscode.window.showInformationMessage("Connected to the GDScript language server.");
-				break;
-			case ClientStatus.DISCONNECTED:
-				this.retry_connect_client();
-				break;
-		}
-	}
-
-	// TODO: move to LSP client?
-	private on_client_status_changed(status: ClientStatus) {
-		let host = get_configuration("lsp.serverHost", "localhost");
-		let port = get_configuration("lsp.serverPort", 6008);
-		switch (status) {
-			case ClientStatus.PENDING:
-				this.connection_status.text = `$(sync) Connecting`;
-				this.connection_status.tooltip = `Connecting to the GDScript language server at ${host}:${port}`;
-				break;
-			case ClientStatus.CONNECTED:
-				this.retry = false;
-				set_context("godotTools.context.connectedToEditor", true);
-				this.connection_status.text = `$(check) Connected`;
-				this.connection_status.tooltip = `Connected to the GDScript language server.`;
-				if (!this.client.started) {
-					this.context.subscriptions.push(this.client.start());
-				}
-				break;
-			case ClientStatus.DISCONNECTED:
-				if (this.retry) {
-					this.connection_status.text = `$(sync) Connecting ` + this.reconnection_attempts;
-					this.connection_status.tooltip = `Connecting to the GDScript language server...`;
-				} else {
-					set_context("godotTools.context.connectedToEditor", false);
-					this.connection_status.text = `$(x) Disconnected`;
-					this.connection_status.tooltip = `Disconnected from the GDScript language server.`;
-				}
-				this.retry = true;
-				break;
-			default:
-				break;
-		}
-	}
-
-	private retry = false;
-
-	// TODO: move to LSP client?
-	private retry_callback() {
-		if (this.retry) {
-			this.retry_connect_client();
-		}
-	}
-
-	// TODO: move to LSP client?
-	private retry_connect_client() {
-		const auto_retry = get_configuration("lsp.autoReconnect.enabled", true);
-		const max_attempts = get_configuration("lsp.autoReconnect.attempts", 10);
-		if (auto_retry && this.reconnection_attempts <= max_attempts) {
-			this.reconnection_attempts++;
-			this.client.connect_to_server();
-			this.connection_status.text = `Connecting ` + this.reconnection_attempts;
-			this.retry = true;
-			return;
-		}
-
-		this.retry = false;
-		this.connection_status.text = `$(x) Disconnected`;
-		this.connection_status.tooltip = `Disconnected from the GDScript language server.`;
-
-		let host = get_configuration("lsp.serverHost", "localhost");
-		let port = get_configuration("lsp.serverPort", 6008);
-		let message = `Couldn't connect to the GDScript language server at ${host}:${port}. Is the Godot editor or language server running?`;
-		vscode.window.showErrorMessage(message, "Open Godot Editor", "Retry", "Ignore").then(item => {
-			if (item == "Retry") {
-				this.connect_to_language_server()
-			} else if (item == "Open Godot Editor") {
-				this.client.status = ClientStatus.PENDING;
-				this.open_workspace_with_editor("-e").then(() => {
-					setTimeout(() => {
-						this.reconnection_attempts = 0;
-						this.client.connect_to_server();
-					}, 10 * 1000);
-				});
 			}
 		});
 	}
